// frontend/src/app/core/models/athlete.modelBD.ts
import { PerformanceMetrics, DailyStatus, BodyComposition, Complaint, LabExam } from "./healthData.model";
import { TimeSeriesDataPoint, BloodPressureDataPoint, CaloricIntakeDataPoint, HormonalLevel, Availability } from "./healthChartData.model";

type levelType = 'Sênior' | 'Júnior'
type statusType = 'Ok' | 'Lesionado' | 'Restringido'

export default interface AthleteBd {
  name: string;
  photoUrl: string;
  age: number;
  weight: number;
  height: number;
  focus: string;
  gender: string;
  level: levelType;
  status: statusType;
  medicalConditions: string[];

  // Dados do Perfil Detalhado
  performanceMetrics?: PerformanceMetrics;
  dailyStatus?: DailyStatus;
  bodyComposition?: BodyComposition;
  complaintHistory?: Complaint[];
  bloodPressureHistory?: BloodPressureDataPoint[];
  weightHistory?: TimeSeriesDataPoint[];
  caloricIntakeHistory?: CaloricIntakeDataPoint[];
  labExams?: LabExam[];
  hormonalLevels?: {
    cortisol: HormonalLevel;
    testosterone: HormonalLevel;
  };
  availabilityLast30Days?: Availability;
}